reanaming commons to lib.js
